requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
brotlipy==0.7.0
cached-property==1.5.2
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.7
cryptography==41.0.1
cycler==0.11.0
flatbuffers==23.5.26
fonttools==4.39.4
frozenlist==1.3.3
gast==0.4.0
google-auth==2.19.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.8.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
joblib==1.2.0
JPype1==1.4.1
keras==2.11.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
konlpy==0.6.0
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
multidict==6.0.4
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
pip==23.1.2
platformdirs==3.5.1
pooch==1.7.0
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pyu2f==0.1.5
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
setuptools==67.7.2
six==1.16.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.1
tensorflow-estimator==2.11.0
termcolor==2.3.0
threadpoolctl==3.1.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.3.4
wheel==0.40.0
wrapt==1.15.0
yarl==1.9.2
zipp==3.15.0
